+++ src/pages/TeamsPage.tsx
@@ -1,5 +1,5 @@
 import Nav from '@/components/common/Nav';
-import SideBar from '@/components/common/SideBar';
+import SideBar from '@/components/common/SideBar/SideBar';
 import TeamBar from '@/components/TeamBar';
 import AnnouncementBoard from '@/components/announcement/AnnouncementBoard';
 import KanbanBoard from '@/components/issue/KanbanBoard';
